add sourcing-from-drupal.md (#8753)
* Update sourcing-from-drupal.md

* minor links and content fixes

* Update sourcing-from-drupal.md

* added config code into a code block

* A few minor edits

Hi! Thanks for creating this doc! It looks great. I changed the word "simple" to "only a few steps" to match with the Gatsby Style Guide and made a few more minor edits.

I'm trying to decide whether the code snippet is useful since now there is a link to the plugin's READme file anyway.

* Changed example site

Since this is now documentation, I changed the example site URL to the source code instead of the site itself. Does that seem like a good idea?

* Update sourcing-from-drupal.md

* Update sourcing-from-drupal.md