fix(docs): add example for pageContext usage within a page (#11580)
## Description

This page explains very well that you can pass `context` to a page automatically created, but it's missing an example of how to use it in a page via `pageContext`. I think it is important to explicitly explain it because the name of the variable is different to the prop name.
I added an example.

## Related Issues
I didn't find any related issues.