Update programmatically-creating-pages.md (#22944)
There is a slight inconsistency in the docs for the final steps of the `Make a template for your posts` section. 

The docs have separated this section into two steps (creating an MDX layout component + adding a GraphQL query) but when showing the completed component code (`src/components/posts-page-layout.js`)  the component is missing some imports and the `<MDXProvider>` component that were present in the first step of this section.