www: Persist filters on Site Showcase in opened modals (#21650)
* fix prev/next link in showcase modal when filters are applied

* pass an empty array as filter if nothing is selected

When we click on a site in showcase list without selecting any filter, page breaks as `filter` variable is passed as undefined but the function is trying to call an array method on it. So passing an empty array if no filter is selected

* fix prev/next link in showcase modal when filters are applied

* pass an empty array as filter if nothing is selected

When we click on a site in showcase list without selecting any filter, page breaks as `filter` variable is passed as undefined but the function is trying to call an array method on it. So passing an empty array if no filter is selected

* fix merge issues

* remove duplicate variables